feat(calltree): move the resizable logic to hook
